pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["BoManev <bomanev@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = {extras = ["all"], version = "^0.103.2"}
sqlmodel = {git = "https://github.com/BoManev/sqlmodel"}
uvicorn = "^0.23.2"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
requests = "^2.31.0"
numpy = "1.26.0"
pandas = "^2.1.3"
# transformers = "4.32.1"
# torch = "2.0.0"
# scikit-learn = "1.2.2"
httpx = "^0.25.1"
# probably breaking changes once 1.0.0
fastapi-events = "0.9.*"
alembic = "^1.12.1"

[tool.poetry.group.dev.dependencies]
Faker = "^19.10.0"
httpx = "^0.25.0"
pytest = "^7.4.2"
pillow = "^10.1.0"
ruff = "^0.1.6"


[tool.poetry.group.rec_engine.dependencies]
numpy = "1.26.0"
transformers = "4.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
api = "api:start"

[tool.ruff]
target-version = "py311"
line-length = 100
indent-width = 4


[tool.ruff.lint.per-file-ignores]
"api/models/__init__.py" = ["F401"]
"test/**" = ["F811"]
"api/models/**" = ["E712"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"